% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merchandiser-package.R
\docType{package}
\name{merchandiser-package}
\alias{merchandiser}
\alias{merchandiser-package}
\title{Tree merchandising from measured stems}
\description{
Cut logs from tree measurements in inches and feet, explicit products, and located defects.
Measure stem dimensions and volume, fit heights and taper, and estimate dry mass and carbon.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://siskiyoubiometrics.com/merchandiser/}
  \item \url{https://github.com/siskiyoubiometrics/merchandiser}
  \item Report bugs at \url{https://github.com/siskiyoubiometrics/merchandiser/issues}
}

}
\author{
\strong{Maintainer}: Hunter Stanke \email{hunter@siskiyoubiometrics.com} [copyright holder]

Authors:
\itemize{
  \item Hunter Stanke \email{hunter@siskiyoubiometrics.com} [copyright holder]
}

}
\keyword{internal}
