% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem-registry.R
\name{has_taper_model}
\alias{has_taper_model}
\title{Check whether taper equations are available}
\usage{
has_taper_model(
  model
)
}
\arguments{
\item{model}{The identifier names the taper model in the registry. Character vector of model
identifiers. Required, with no default.}
}
\value{
A logical vector, TRUE where a registered taper model is available.
}
\description{
Check whether taper equations are available
}
\examples{
## Check availability of the example trees' models.
has_taper_model(model = example_trees$model)
}
