% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem-divisions.R
\docType{data}
\name{nsvb_division_polygons}
\alias{nsvb_division_polygons}
\title{Ecological division boundaries}
\format{
A data frame with division (numeric division code), ring_id (polygon
ring identifier), hole (logical interior ring indicator), x (longitude,
degrees), and y (latitude, degrees).
}
\usage{
nsvb_division_polygons
}
\description{
Boundary coordinates support ecological division lookup.
}
\keyword{datasets}
