% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem-registry.R
\name{unregister_taper_model}
\alias{unregister_taper_model}
\title{Remove a local equation from the current analysis}
\usage{
unregister_taper_model(
  model
)
}
\arguments{
\item{model}{The identifier names the taper model in the registry. Character scalar model
identifier. Required, with no default.}
}
\value{
TRUE, invisibly, after removing the private model.
}
\description{
Remove a local equation from the current analysis
}
\examples{
## Copy an example tree's model
local_model <- get_taper_model(model = example_trees$model[1])

## Name the local copy
local_model$id <- 'example.remove'

## Register the copy
register_taper_model(x = local_model)

## Remove and report the local registration
print(unregister_taper_model(model = local_model$id))
}
